monaco_demo/requirements.txt
git+https://github.com/reflex-dev/reflex.git
reflex-monaco
